Fix object_ids call to be less intense
Previously, if you had set highlights, then object_ids() would call
raw() and raw() would build a query that included a request for highlighted
stuff and then you'd get a ton of JSON back when all you wanted were
ids.

This code is specific to SUMO and uses code in elasticutils that is
also specific to SUMO. Given that, I'm abusing the freedom to do a
hacky fix where we gimp _build_highlight, call raw(), then ungimp it.
This makes sure no highlight stuff shows up in the query and thus
we don't get highlighted stuff in the returned JSON (and ES doesn't
do all that highlighting work).